修复次日回访结构性抓不到+回填不看交易日+内置AI开放全能问答
- advisor.py/_backfill_due_advice: watchlist回访窗口从1天改0.9天——
  cron每天固定同一时刻启动，watchlist判断排在main()最后，落库时间
  总比当天cron启动晚10-20分钟，导致"距上次cron启动正好24小时"这个
  cutoff结构性地永远卡在昨天watchlist落库时间之前，实测8-28这批49条
  在真实cutoff下命中数是0，不是偶尔慢是每天必然卡住
- 回填价格等于入场价时不记review_price，大概率是当天没开盘（周末/
  节假日回填取到的是上一交易日收盘价）——8-28那批周六回填出的31条
  review_price=price_at_advice污染了回测，把胜率拉到12%~26%
  （正常该在50%上下），现在这类情况留NULL不硬记0%收益
- app.py._backfill_due_reviews同样的equal-price保护，另外把它的
  超时从3秒调到10秒（比自己docstring里写的实测10秒延迟还短，等于
  经常一条都补不上）
- position/screen回填limit从默认20提到200，加速清理积压
  （不影响总耗时，deadline仍是固定60秒不是按条数算）
- 删除113条手动测试污染的position记录（8-21日6条+8-28日107条，
  症状是当天没有的股票代码），跟真实持仓表(00700/600519/SKHY)对齐
- assistant.py: 系统提示词放开"全能金融专家"定位——网站自己的数据
  之外的股票/公司/新闻/宏观知识性问题不再受"看不到就说看不到"那条
  规则约束；新增三个工具(实时行情/财务摘要/新闻搜索)，千问自己判断
  要不要调，覆盖用户随口问的任意股票，不限于首页排行榜/持仓里已有的
  那几十支——买卖建议类问题的回避规则不变，这条不分是否网站收录都成立

+++ tracker.py
@@ -710,7 +710,7 @@ def log_advice(
 
 
 def get_due_for_advice_review(
-    email: str, min_age_days: int = 7, limit: int = 20, source: str | None = None,
+    email: str, min_age_days: float = 7, limit: int = 20, source: str | None = None,
 ) -> list[dict]:
     """同 get_due_for_review 的逻辑，找出该回填实际价格的历史建议（只看当前用户）。
